patched init.sh

+++ init.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-echo -n "Do you want to fully initialise (or on AWS)(y/n)?: "
+echo -n "Do you want to fully initialise (or on AWS)(y/n)? "
 read ch
 
-if [ "$ch" -eq "y" ] || [ "$ch" -eq "yes" ]; then
+if [ "$ch" = "y" ] || [ "$ch" = "yes" ]; then
     # Update the package list
     sudo apt-get update -y
 
@@ -30,24 +30,34 @@ gcc --version
 echo "G++ version:"
 g++ --version
 
-echo "The Python virtual environment is created and activated.\e[0m"
+echo "\e[0mIf any of the above commands resulted in an error, exit and restart script with initialisation on."
+
+if [ -d ".env" ]; then
+    echo -n "Existing environment found, do you want to delete it? (y/n)? "
+    read ch
+    if [ "$ch" = "y" ] || [ "$ch" = "yes" ]; then
+        echo "Deleting existing environment."
+        rm -rf .env
+    fi
+fi
 
 # Create a Python virtual environment in the current directory
 python3 -m venv .env
 
 # Activate the virtual environment
 source .env/bin/activate
+echo "\e[32mThe Python virtual environment is created and activated.\e[0m"
 
 echo "\e[33mPython requirements will be fetched and installed now.\e[0m"
 
 # Install packages from requirements.txt if it exists
+pip install --upgrade pip
 if [ -f "requirements.txt" ]; then
     pip install -r requirements.txt
+    echo "\e[32mThe Python requirements are installed.\e[0m"
 else
     echo "\e[31mrequirements.txt not found, skipping package installation.\e[0m"
 fi
-
-echo "\e[32mThe Python requirements are installed.\e[0m"
 
 # compiling the shared library for evaluate cards
 echo "\e[33mShared library for evaluate cards will be compiled now.\e[0m"
@@ -58,8 +68,10 @@ cd build
 cmake ..
 make pheval
 cd ..
-mv build/libpheval.0.6.0.dylib build/libpheval.so.0.6.0
+if [ -f "build/libpheval.0.6.0.dylib" ]; then
+    mv build/libpheval.0.6.0.dylib build/libpheval.so.0.6.0
+fi
 g++ -fPIC --shared -std=c++11 -I include/ wrapper.cpp build/libpheval.so.0.6.0 -o wrapper.so
 cd ..
 
-echo "\e[32mShared library for evaluate cards compiled.\e[0m"
+echo "\e[33mProcess complete, check above log for errors and restart if needed.\e[0m"
